Simplify CFX_StandardFont standard font lookup
* Replace GetStandardFontName(ByteString*), which mutated its input
  argument, with a non-mutating GetStandardFontIndex() method.
* Callers that need the canonical standard font name must now explicitly
  call GetCanonicalFontName() rather than relying on a side-effect.
* Update affected call sites across fpdfapi/ and fxge/.

TAG=agy
CONV=63532648-374f-4a87-acdf-e5aebbaa70a3

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/150690
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fpdfapi/font/cpdf_stockfontarray.cpp
@@ -29,13 +29,13 @@ CPDF_StockFontArray::~CPDF_StockFontArray() {
 
 RetainPtr<CPDF_Font> CPDF_StockFontArray::GetFont(
     CFX_StandardFont::StandardFont index) const {
-  CHECK_LT(index, std::size(stock_fonts_));
+  CHECK_LT(static_cast<size_t>(index), std::size(stock_fonts_));
   return stock_fonts_[index];
 }
 
 void CPDF_StockFontArray::SetFont(CFX_StandardFont::StandardFont index,
                                   RetainPtr<CPDF_Font> font) {
-  if (index < std::size(stock_fonts_)) {
+  if (static_cast<size_t>(index) < std::size(stock_fonts_)) {
     stock_fonts_[index] = std::move(font);
   }
 }
